pyproject.toml
[project]
name = "voise"
version = "0.1.0"
description = "Real-time AI voice assistant powered by Gemini Live API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "opencv-python",
    "pyaudio",
    "pillow",
    "mss",
    "python-dotenv",
    "pyyaml>=6.0.3",
]

[tool.uv]
managed = true
